+++ cartridges/int_marketpay_headless/cartridge/scripts/hooks/basketPaymentHooks.js
@@ -84,7 +84,6 @@ exports.modifyGETResponse_v2 = function (basket, paymentMethodResultResponse) {
 
             // If mapping exists, add terminal info to the method
             if (terminalMapping !== null) {
-
                 // Match terminalMapping.name with marketPayPaymentMethods metadata.terminalName
                 if (marketPayPaymentMethods && marketPayPaymentMethods.methods && marketPayPaymentMethods.methods.length > 0) {
                     for (var k = 0; k < marketPayPaymentMethods.methods.length; k++) {
@@ -94,14 +93,12 @@ exports.modifyGETResponse_v2 = function (basket, paymentMethodResultResponse) {
                             delete paymentMethod.onInitiatePayment;
 
                             method.c_marketPay = {
-                                //sessionId: marketPayTokenAndSession.sessionId,                            
                                 paymentMethod: paymentMethod
                             }
-                            break;
+                            return true;
                         }
                     }
                 }
-                return true;
             }
             return false;
         });
@@ -129,7 +126,6 @@ exports.afterPOST = function (basket, paymentInstrument) {
         // Get sessionId from Custom Object MarketPayData for this user
         var CustomObjectMgr = require('dw/object/CustomObjectMgr');
         var marketPayDataObj = CustomObjectMgr.getCustomObject('MarketPayData', basket.customer.ID);
-        var marketPayToken = marketPayDataObj ? marketPayDataObj.custom.token : null;
         var marketPaySessionId = marketPayDataObj ? marketPayDataObj.custom.sessionID : null;
         var marketPayPaymentMethods = marketPayDataObj ? marketPayDataObj.custom.paymentMethods : null;
         var onInitiatePaymentURL = marketPayDataHelper.getOnInitiatePaymentURL(paymentInstrumentRequest.c_marketPayPaymentMethodID, marketPayPaymentMethods);